requirements.txt
tensorflow
torch==2.0.1 
torchvision==0.15.2
transformers==4.34.0
scikit-learn==1.3.0
Pillow==9.4.0
numpy==1.24.2
matplotlib==3.7.1